remove connected apps
