+++ db/migrate/20220309153801_add_argo_token_to_user.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Add argo token to users table
+class AddArgoTokenToUser < ActiveRecord::Migration[7.0]
+  def change
+    add_column :users, :argo_token, :string
+  end
+end
