--- test/fixtures/connected_apps.yml
@@ -1,11 +1,0 @@
-# Read about fixtures at https://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
-
-one:
-  user_id: 3
-  app_id: 2
-  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJqdGkiOiI3MzUzYmJkYi1kYTA1LTQxYjAtOGY5MC0zMzUxZDRhNDE3YjEiLCJpYXQiOjE2NDY3NzQ1NTksImlzcyI6ImFyZ29jZCIsIm5iZiI6MTY0Njc3NDU1OSwic3ViIjoidGVzdF91c2VyOmFwaUtleSJ9.kekkvIzuNOU2FC2dgFLUHyrbI8aScrQxE7Dz5GzoG20
-
-two:
-  user_id: 3
-  app_id: 1
-  token: MyString
