--- test/models/connected_app_test.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require 'test_helper'
-
-class ConnectedAppTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end
